// src/service/file/schemas/apiResponseListSmartFolderItemResult.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * File API
 * isd / kinder board api doc
 * OpenAPI spec version: v1.1.0
 */
import type { SmartFolderItemResult } from './smartFolderItemResult';

export interface ApiResponseListSmartFolderItemResult {
  status: number;
  result?: SmartFolderItemResult[];
  timestamp?: string;
  error?: string;
  debug?: string;
}